Update student controller download function
Update lecturer controller file structure for download function

Update admin controller to add module coodinator functions

Update compiler.cs to retrieve correct application

Add grade class in grading folder

Update Sandboxer to retrieve and store data

Update database model under DAL folder in Models folder

Update ManageLecturerViewModel to include ModuleCoordinator variable

Update UpdateLecturerViewModel to include required variables for view

Update addLecturerViewModel to include validation variables

Add Result class to store student result

Add submissionSolution class to retrieve student solutions and results

Update admin views

Update lecturer views

Update student views

+++ SPade/SPade/Models/HangfireData.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
+using SPade.ViewModels.Student;
 
 namespace SPade.Models
 {
@@ -11,6 +12,6 @@ namespace SPade.Models
         public DateTime SucceededAt { get; set; }
         public int PerformanceDuration { get; set; }
         public int Latency { get; set; }
-        public decimal Result { get; set; }
+        public int Result { get; set; }
     }
 }
